+++ src/store/modules/page.js
@@ -3,23 +3,28 @@ import * as types from '../mutation-types'
 // initial state
 const state = {
   url: '',
+  content: [],
 }
 
 // getters
 const getters = {
-  url: state => state.url,
+  pageUrl: state => state.url,
+  pageContent: state => state.content,
 }
 
 // actions
 const actions = {
   updateUrl ({ commit }, url) {
-    commit(types.UPDATE_URL, { url })
+    commit(types.PAGE_UPDATE_URL, { url })
   },
 }
 
 // mutations
 const mutations = {
-  [types.UPDATE_URL] (state, { url }) {
+  [types.PAGE_UPDATE_CONTENT] (state, { content }) {
+    state.content = content
+  },
+  [types.PAGE_UPDATE_URL] (state, { url }) {
     state.url = url
   },
 }